mxe-test/requirements.txt
robotframework
robotframework-browser
robotframework-requests
robotframework-seleniumlibrary
robotframework-jsonlibrary
robotframework-kubelibrary
robotframework-xvfb
robotframework-jmeterlibrary
robotframework-pabot
robotframework-jupyterlibrary
twine
pandas
retry
python-keycloak==0.25.0
requests
urllib3
deepdiff
pexpect
yq